+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ## [2026-09-09]
 - Admins must specify which video within content age user receives
 
+## [2026-09-08]
+- Send a different "activities not suitable" auto-response depending on the child's age in months
+
 ## [2026-09-02]
 - Add cohorts to users to differentiate batches